libv.net: Initial network skeleton and net ts adoption

+++ include/libv/meta/derived_level.hpp
@@ -2,21 +2,18 @@
 
 #pragma once
 
-// std
-#include <cstddef>
-
 
 namespace libv {
 namespace meta {
 
 // -------------------------------------------------------------------------------------------------
 
-template <size_t N> struct derived_level : derived_level<N - 1> {
-	static constexpr size_t value = N;
+template <int N> struct derived_level : derived_level<N - 1> {
+	static constexpr int value = N;
 };
 
 template <> struct derived_level<0> {
-	static constexpr size_t value = 0;
+	static constexpr int value = 0;
 };
 
 using derived_top = derived_level<255>;
